useFetchData refactoring
