+++ src/components/Loader.styled.tsx
