+++ src/components/NoContent.tsx
@@ -1,0 +1,10 @@
+import React from "react"
+import { Container } from "./basic/Container"
+
+export function NoContent() {
+  return (
+    <Container>
+      <h1>No content provided from API...</h1>
+    </Container>
+  )
+}
