+++ src/pages/HomePage.styled.tsx
@@ -1,0 +1,7 @@
+import styled from "styled-components"
+
+export const ExperimentsSection = styled.section`
+  display: flex;
+  flex-wrap: wrap;
+  gap: 1rem;
+`
