+++ src/pages/HomePage.tsx
@@ -1,13 +1,14 @@
 import React from "react"
 import { useFetchData } from "../hooks/useFetchData"
 import { apiClient } from "../api/apiClient"
-import { FullscreenLoaderOverlay } from "../components/FullscreenLoaderOverlay"
+import { Loader } from "../components/Loader"
+import { NoContent } from "../components/NoContent"
 
 export default function HomePage() {
   const { data, isLoading } = useFetchData(apiClient.experiments.getAll)
-
+  console.log({ data })
   if (!data) {
-    return isLoading ? <FullscreenLoaderOverlay /> : null
+    return isLoading ? <Loader /> : <NoContent />
   }
 
   return (
